fixup formatting of catmull-rom spline article
When I converted this article, I used the pandoc markdown format instead
of commonmark. This behaves differently from kramdown that we use to
render the article, leading to markup mistakes.

This commit tries to re-convert the article, but keep all changes done
on top. The result is a bit of a mix between reflowing (mostly
re-wrapping lines that was manually fixed up) and fixing markup mistakes
(mainly switching hard line-feed syntax).

I'm sorry that it's a bit hard to read the resulting diff, but this just
seems easier than doing it all by hand.