set max-width
In typograph, it's generally considered that overly long lines of text
take longer to read. So let's try to combat that a bit, and cap the max
size of the main-container.

We can't really do this based on proper scientific data, because we use
proportional fonts, and there's some extra margins etc involved. But we
can do something approximate instead.

The exact value is based on taste and trying to make sure text and
layout doesn't break too bad in extreme but not too unrealistic sizes.
It's not perfect, but probably better than nothing.